Provide initializers #62
