+++ src/packages/piral-cli/src/index.ts
@@ -1,2 +1,3 @@
 export * from './helpers';
 export * from './commands';
+export * from './questionnaire';
